+++ __mocks__/lucide-react-native.js
@@ -145,4 +145,5 @@ module.exports = {
   Layout: createIconMock('Layout'),
   Sparkles: createIconMock('Sparkles'),
   Flame: createIconMock('Flame'),
+  QrCode: createIconMock('QrCode'),
 };